backend/requirements.txt
flask
flask-cors
pandas
pytz
mysql-connector-python
sqlalchemy
pymysql